Puzzle Update 1
Add "PedroTestLevel" Which is a copy of the "PlayerMovementTest"
Rename (Previous) "PedroTestLevel" to "PauseMenuTest"

Pedro test level contains a trigger box between a cylinder and a box that will turn a light on if the player enters and turn it off if the player exits. Trigger box only responds to player entering/ exiting.