Working on Getting Child Comps
After a while of debugging and research I was unable to find anything about how to retieve the children components of the actor. The Root Comp is showing up but its not returning any of it's children comps. So I commented out to prevent any issues from arising during testing.